+++ src/main/java/in/koreatech/koin/domain/ownershop/controller/OwnerEventController.java
@@ -5,18 +5,7 @@ import static in.koreatech.koin.domain.user.model.UserType.OWNER;
 import in.koreatech.koin.domain.ownershop.dto.CreateEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.ModifyEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.OwnerShopEventsResponse;
-import in.koreatech.koin.domain.ownershop.dto.OwnerShopsRequest;
-import in.koreatech.koin.domain.ownershop.dto.OwnerShopsResponse;
-import in.koreatech.koin.domain.ownershop.service.OwnerShopService;
-import in.koreatech.koin.domain.shop.dto.menu.CreateCategoryRequest;
-import in.koreatech.koin.domain.shop.dto.menu.CreateMenuRequest;
-import in.koreatech.koin.domain.shop.dto.menu.MenuCategoriesResponse;
-import in.koreatech.koin.domain.shop.dto.menu.MenuDetailResponse;
-import in.koreatech.koin.domain.shop.dto.menu.ModifyCategoryRequest;
-import in.koreatech.koin.domain.shop.dto.menu.ModifyMenuRequest;
-import in.koreatech.koin.domain.shop.dto.menu.ShopMenuResponse;
-import in.koreatech.koin.domain.shop.dto.shop.ModifyShopRequest;
-import in.koreatech.koin.domain.shop.dto.shop.ShopResponse;
+import in.koreatech.koin.domain.ownershop.service.OwnerEventService;
 import in.koreatech.koin.global.auth.Auth;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -29,13 +18,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequiredArgsConstructor
 public class OwnerEventController implements OwnerEventApi {
 
-    private final OwnerShopService ownerShopService;
+    private final OwnerEventService ownerEventService;
 
     @PostMapping("/owner/shops/{id}/event")
     public ResponseEntity<Void> createShopEvent(
@@ -43,7 +31,7 @@ public class OwnerEventController implements OwnerEventApi {
         @PathVariable("id") Integer shopId,
         @RequestBody @Valid CreateEventRequest shopEventRequest
     ) {
-        ownerShopService.createEvent(ownerId, shopId, shopEventRequest);
+        ownerEventService.createEvent(ownerId, shopId, shopEventRequest);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
@@ -54,7 +42,7 @@ public class OwnerEventController implements OwnerEventApi {
         @PathVariable("eventId") Integer eventId,
         @RequestBody @Valid ModifyEventRequest modifyEventRequest
     ) {
-        ownerShopService.modifyEvent(ownerId, shopId, eventId, modifyEventRequest);
+        ownerEventService.modifyEvent(ownerId, shopId, eventId, modifyEventRequest);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
@@ -64,7 +52,7 @@ public class OwnerEventController implements OwnerEventApi {
         @PathVariable("shopId") Integer shopId,
         @PathVariable("eventId") Integer eventId
     ) {
-        ownerShopService.deleteEvent(ownerId, shopId, eventId);
+        ownerEventService.deleteEvent(ownerId, shopId, eventId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
@@ -73,7 +61,7 @@ public class OwnerEventController implements OwnerEventApi {
         @Auth(permit = {OWNER}) Integer ownerId,
         @PathVariable("shopId") Integer shopId
     ) {
-        OwnerShopEventsResponse shopEventsResponse = ownerShopService.getShopEvent(shopId, ownerId);
+        OwnerShopEventsResponse shopEventsResponse = ownerEventService.getShopEvent(shopId, ownerId);
         return ResponseEntity.ok(shopEventsResponse);
     }
 }
